Add files via upload
This was my final project for Financial Econometrics. We utilized data from the VIX and TYVIX to model the relationship between the two using different models. We tested for the data's stationarity and eventually found good fits for an ARIMA, ARCH, and GARCH models. After that, we tested to see if the two series could be modeled via a Vector Autoregression model and visualized the Impulse Response Function for such a model. Finally, we used the ARCH/GARCH specifications to forecast the VIX's volatility and mean, generated normal distributions for each forecasted day's expected volatility and mean, and built a Monte Carlo simulation.